requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
openpyxl==3.1.2
PyMuPDF==1.23.8
mammoth==1.6.0
pandas==2.1.4
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
